(PC-4517) Remove 'get' command
The command imports code from `tests.conftest`, which will soon not
work anymore: when we remove the PYTHONPATH override, the "tests"
package will not be installed and will not be importable
anymore (except from the tests themselves, of course).

Since the command does not seem to be used, let's remove it.