(PC-4715) Reorganize booking-related code
A version of the "clean architecture" was partially implemented in
this project for booking-related code. It has been found to be more
complex than necessary.

This commit aims at implementing the very same features with a more
simple, MVC-like design. Most notably, there is no distinction between
the "database model" and the "domain model" (and the latter has been
removed).

Most of the code has been moved to a new `pcapi.core.bookings`
sub-package that contains the following modules:

- models: this module defines booking-related SQLAlchemy-based
  models. Only `Booking`, for now.

- api: where the logic is implemented. All actions that change data
  should go through a function defined in this module.

- validation: implements checks that are to be used by functions of
  the "api" module, mostly.

- exceptions: contains all booking-related exceptions.

- repository: contains functions that access the database without
  changing data. In other words, these make custom SELECT queries.

- helpers: booking-specific utility functions.

Routes and scripts have been adapted to use the functions defined in
`pcapi.core.bookings.api`.

Tests have been moved to "tests/core/bookings" and sometimes partially
rewritten. The intent of the tests have been kept and the coverage
should be similar to what we had before. Also, we use this as an
opportunity to use Factory Boy, which makes for more concise tests.

And a lot of code has been removed in the process. :)

Other simplifications can be made later, in another commit, because
this one is already pretty large.