(PC-4233) Activate reCaptcha v3 challenge for ID CHECK
Google's reCaptcha docs can be found at https://developers.google.com/recaptcha/docs/v3

This change allows us to handle the different errors that the Google API can return,
and implements stricter control over the reCaptcha v3, via the score and a check on the action.

The minimum required score is 0.5 by default, and it can be tuned via a new env variable.