(PC-5077) payment: Send a zipped CSV file for payment details
The "payment details" CSV file that we send by e-mail is getting too
large. The last (2020-11-03) CSV file is 12 Mb. Once base64-encoded,
it is around 16.6 Mb, which is above the 15 Mb limit set by Mailjet.
When we tried to send the e-mail with that file, we got the following
error when POSTing to Mailjet API:

    Connection aborted: BrokenPipeError(32, 'Broken pipe')

It's not clear whether the error happens because of the size limit: I
could not find any reference in Mailjet documentation nor elsewhere.
But it's a good guess. :)

At least, I have been able to successfully send a zipped CSV file,
which weighted a bit more than 1 Mb.