bookings: Check unicity of booking token before inserting into the database
We used to generate a random token and hope that it was not already
used. Lately, we were getting relatively frequent UNIQUE constraint
error because of that.

Now we check that the randomly-generated token is not used before
INSERTing it in the database. Theoretically, we could still have two
concurrent bookings that generate the same token at the same time, but
it is very unlikely. A workaround would be to acquire a lock before
(or let the database generate the token) but I don't think it's worth.