(PC-5536) utils: Increase the number of permutations of the random token generator
The implementation was somewhat hard to read and had "only"
255^(length / 2) possible permutations (according to Cyril who is
better at math than me). For the default 6-character tokens, that's
around 16 millions permutations.

The new implementation is a lot clearer, and you can easily tell that
it can generate 30^length permutations. For 6-character tokens, that's
around 729.000.000 permutations, which should be enough for a while.

The alphabet has been chosen to keep the same scheme as before (based
on `humanize()`): all letters and digits of Base 32 (A-Z and 2-7)
except letters O and I to avoid misreadings.