(PC-5613) alembic: Restore event_is_in_less_than_10_days SQL function
The function was deleted in 55fe8e716f6555d210744e3fedc2a8ffb034f737.

... but it is still used in staging and prod because the definition of
the `get_offer_score` SQL function has not been updated on these
environments. In other words, on staging and prod, we still need the
`event_is_in_less_than_10_days` SQL function.

We should create a new migration that updates `get_offer_score` on
staging and prod, but that warrants a thorough validation, and I am
not willing to change recommendation-related code in a patch release.