testing: Fix override_features and don't use it on test classes
Move `override_features` to the testing module (so that all
testing-related helpers are defined in the same place). We now use the
new `TestContextDecorator` base class to implement the decorator
(see previous commit for further details). As indicated in the
previous commit, the former implementation hid decorated classes from
pytest: tests were not run.

Unfortunately, the new implementation does not play well with the
`db_session` fixture from pytest_flask_sqlalchemy when applied on a
class (see comments). We must apply this decorator on each method of
the class. Until a solution is found, we make the decorator fail when
used on a class: at least, no test will be ignored from now on.