(PC-6807) : Fix null prices for existing stocks
This issue was tricky. Null prices were fixed BUT for existing stocks
there was some strange integrity error risen when the error log was
added in the DB.
It turns out that the stock is being modified then validated.
The validation fails and the backend inserts a message in the DB
about the failure.
HOWEVER when inserting the message a commit happens and since the
stock was modified SQLAlchemy assumes the stock is meant to be
commited.
Because the stock did not meet the validation requirements it also
fails the DB constraints (null price is not allowed) and thus an
integrity is risen.

I HATE implicit actions with SQLAlchemy

So in order to prevent thos cases, I need to EXPLICITLY remove the
stock from the session by expiring it.