(PC-6752) Let get_wallet_balance SQL function return a negative number
The `check_booking` SQL function raises the `insufficientFunds` error
when `get_wallet_balance() < 0`. This never happens if we use
`GREATEST(0, xxx)`. As such, it would not prevent a user to book again
with an expired deposit if they had booked something in the past.

What we really want is to avoid exposing a negative deposit value, so
we'd rather do it in `User` properties.