(PC-5892) Stop using UserOfferer.rights before its removal
This attribute is not used anymore. On production, all instances of
UserOfferer are either 'editor' or 'admin'. The code only grants the
'editor' role. All in all, we are only interested in knowing whether a
User has an associated UserOfferer for a given Offerer.

Changes:

- `UserOfferer.rights` is not used anymore. The field is already
   nullable. We will remove it (through a migration) after the
   deployment of v123.

- `User.hasRights()` has been renamed as `has_access()` and its
  signature has been adapted.

- `Offerer.giveRights()` has been renamed as `grant_access()` and its
  signature has been adapted.

- `ensure_current_user_has_rights()` has been renamed as
  `check_user_has_access_to_offerer()` for clarity (with the "check_"
  prefix for homogeneity).

Future: the field is already nullable and can removed once v123 is
deployed.