(PC-7924) Make sure that all loggers use our custom record factory
With the previous implementation, loggers that were created before we
called `install_logging()` would not use our custom record factory. As
such, the `extra` property in the JSON log was empty. It occurred on
some (but not all) of our own loggers, and depended on the order in
which modules were imported.

An alternative would be to call `install_logging()` first on each
entry point module of the app (`flask_app.app`, `workers.worker`, each
cron module, etc.) but it's tedious.

The easy solution seems to be a monkey patch. Will I regret it?