(PC-7783) bookings: Make mark_as_used a no-op if booking is already used
Some bookings will soon be automatically marked as used when created.
When an offerer wants to mark a booking as used (through our HTTP
API), they will not necessarily check first whether the booking has
already been marked as used. Also, they cannot know if the booking was
automatically marked as used or if it was created before we changed
the rule. So they are likely to just mark the booking as used: this
operation should hence be a no-op and it should not fail.