(PC-7963) : Add postalCode to the native app account creation
`postalCode` will be required for native app account creation
as long as the product is restricted to some area.
Also note that it is currently flagged as optional - for now -
to avoid breaking account creation. Mandatory flag will be
added in a second step depending on the value of the
`WHOLE_FRANCE_OPENING` flag.