(PC-8249) Do not create offer with empty stock in provider synchronization
When we synchronize offers from a provider, we may receive data for an
offer with no available quantity that is unknown to us. In that case,
we should not create an offer. Adding this offer would be useless,
because it could not be booked and would not even appear in search
results.