providers: Log if we override a specific price when provider price is missing
It may be possible that a provider first sends a specific price (that
is different from the product price), and then sends an update without
any price. Does that case really happen? If so, we're not sure what we
should do. Currently, we override the price with the (generic) product
price. But maybe we should keep the original specific stock price?