(PC-9311) admin: Fix XSS vulnerabilities in offer validation view
Also, use `Markup.format()` with explicitly escaped arguments
everywhere. This is not strictly needed (since `Markup.format()`
already escapes arguments), but this could help us not to introduce a
similar vulnerability by copy-pasting code and then trying to be
clever with an f-string.

Credit goes to Aethlios who responsibly disclosed this vulnerability.