(PC-8776) Remove old provider synchronization code
FNAC_SYNCHRONIZATION_V2 has been enabled for months;
`SYNCHRONIZE_LIBRAIRES` and `SYNCHRONIZE_TITELIVE` have been disabled
as well. We now use the code from `core.providers.api` to synchronize
*stocks* with all providers except Allociné. A lot of code can thus be
removed.

We still need the `LocalProvider` class for the synchronization of
stocks from Allociné, as well as the synchronization of *products*.